+++ CHANGELOG.md
@@ -20,6 +20,8 @@
     * Remove `vaultVenmo` on `DropInRequest` in favor of setting `shouldVault` on `VenmoRequest`
   * DropInPaymentMethodType
     * Rename `DropInPaymentMethodType.GOOGLE_PAYMENT` to `DropInPaymentMethodType.GOOGLE_PAY`
+  * DropInRequest
+    * Remove `requestThreeDSecureVerification` from `DropInRequest`
 
 ## 6.0.0-beta2